Version 0.1 of the password generator.
The generator takes 12 characters of b62 encoded string from a sha1 hash. It
is like supergenpass except that uses md5, and also only uses b16 thus causing
more hash collisions since it is only the 20 most significant bits. This uses
the 64 most significant bits. Thus making it a 1/(2^63) chance of a collision
compared to 1/(2^19) with md5.