Passgen 1.1, modified the way passwords are displayed. Using a new system.
It has been modified so that all numbers save one are shifted to th end of the
string, and that the first alpha character is capatilized, instead of just
being lowercase. This should allow it to get by/be ok for 99.9% of all
password policies out there. I did this so that I could easily type the
passwords on a game console, since there is long runs of similar characters
thus it makes it much much easier to write the data out.