updated awesome_gen_pass.js
Changed back other items within it and also fixed up a bug that was introduced by me getting confused with my own variables when I went in and tried to make them flow properly. Now the system is fixed and will still generate the same variables that it had previously.