Updated Awesome Pass Gen Salt generation & Password Generation
The salt Generation portion of the algorithm has been updated to include a higher number of rounds for 
the password Generation. Further the main password generation algorithm has been updated to increase the output length for the passwords so that they are stronger for the increased security from an attacker who does not already have the cleartext of the password on the site that is hacked without knowing the algorithm.