Updated todo with the new requirements of VCCS
VCCS has some extra requirements that are not previously seen out in the wild to make sure that the generated passwords also pass these requirements I'll have to update the algorithm to take this into account. It is already including an uppercase, a lowercase, a number, and a special character within the generated password but vccs is being more strict it seems. Since I cannot do first/last names within the passwords I'll have to rely on doing substring splicing for the generator.