finally got the testing program completed.
Now there is a set of known good hashes, salts, finalized/simplified hashes to test
against in the future in the event of a change to the core algorithm to make sure that
it always work each time along the way. The overall strength needs to be updated to be
the new values so it's truly inline with the real code.

There are 64 hashes for each of the 3 types of generations for the algorithm. They will
only have to be changed if the core algorithm is updated to a new strength or geeneration
algorithm then all 3 will be updated and each will be moved one level lower.